remove warnings

+++ tt/util/time.h
@@ -18,7 +18,7 @@ public:
     std::string getMessage() const { return m_message; }
 
     float getElapsedMSec() const {
-        return std::chrono::duration_cast<std::chrono::milliseconds>(m_end-m_start).count();
+        return static_cast<float>(std::chrono::duration_cast<std::chrono::milliseconds>(m_end-m_start).count());
     }
 
 private:
